+++ FoodTripz.csproj
@@ -165,12 +165,11 @@
     <Content Include="Content\bootstrap.css.map" />
     <Content Include="Content\bootstrap-theme.min.css.map" />
     <Content Include="Content\bootstrap-theme.css.map" />
-    <Content Include="AdminHomePage.aspx" />
+    <Content Include="AdministratorHomePage.aspx" />
     <None Include="Scripts\jquery-3.4.1.intellisense.js" />
     <Content Include="Login.aspx">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
-    <Content Include="AdministratorHomePage.aspx" />
     <Content Include="PublicDonorPage.aspx" />
     <Content Include="RecipientHomePage.aspx" />
     <Content Include="RecipientRegisteration.aspx" />
@@ -265,13 +264,6 @@
       <DependentUpon>Login.aspx</DependentUpon>
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Compile>
-    <Compile Include="AdministratorHomePage.aspx.cs">
-      <DependentUpon>AdministratorHomePage.aspx</DependentUpon>
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
-    <Compile Include="AdministratorHomePage.aspx.designer.cs">
-      <DependentUpon>AdministratorHomePage.aspx</DependentUpon>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PublicDonorPage.aspx.cs">
       <DependentUpon>PublicDonorPage.aspx</DependentUpon>
@@ -322,12 +314,12 @@
     <Compile Include="ViewSwitcher.ascx.designer.cs">
       <DependentUpon>ViewSwitcher.ascx</DependentUpon>
     </Compile>
-    <Compile Include="AdminHomePage.aspx.cs">
-      <DependentUpon>AdminHomePage.aspx</DependentUpon>
+    <Compile Include="AdministratorHomePage.aspx.cs">
+      <DependentUpon>AdministratorHomePage.aspx</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
-    <Compile Include="AdminHomePage.aspx.designer.cs">
-      <DependentUpon>AdminHomePage.aspx</DependentUpon>
+    <Compile Include="AdministratorHomePage.aspx.designer.cs">
+      <DependentUpon>AdministratorHomePage.aspx</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup />
